Created new auth controller and route with the login function; Deleted the login function from user controller and the route on user route; login route was "api/user/login" and now is "api/auth/login";
