+++ app.js
@@ -4,6 +4,7 @@ const bodyParser = require("body-parser");
 const mongoose = require('mongoose');
 
 const userRoutes = require('./routes/user');
+const authRoutes = require('./routes/auth');
 const notificationRoutes = require('./routes/notification');
 
 const app = express();
@@ -37,6 +38,7 @@ app.use((req, res, next) => {
 });
 
 app.use("/api/user", userRoutes);
+app.use("/api/auth", authRoutes);
 app.use("/api/notification", notificationRoutes);
 
 module.exports = app;
