+++ routes/auth.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const AuthController = require('../controllers/auth');
+
+const router = express.Router();
+
+router.post("/login", AuthController.userLogin);
+
+module.exports = router;
